Rename IResource to Resource
