+++ experimental/packages/otlp-transformer/src/trace/internal.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 import type { Link } from '@opentelemetry/api';
-import { IResource } from '@opentelemetry/resources';
+import { Resource } from '@opentelemetry/resources';
 import type { ReadableSpan, TimedEvent } from '@opentelemetry/sdk-trace-base';
 import type { Encoder } from '../common/utils';
 import {
@@ -112,7 +112,7 @@ export function createExportTraceServiceRequest(
 }
 
 function createResourceMap(readableSpans: ReadableSpan[]) {
-  const resourceMap: Map<IResource, Map<string, ReadableSpan[]>> = new Map();
+  const resourceMap: Map<Resource, Map<string, ReadableSpan[]>> = new Map();
   for (const record of readableSpans) {
     let ilsMap = resourceMap.get(record.resource);
