+++ packages/opentelemetry-resources/src/detect-resources.ts
@@ -15,8 +15,8 @@
  */
 
 import { diag } from '@opentelemetry/api';
-import { IResource } from './IResource';
-import { EMPTY_RESOURCE, resourceFromDetectedResource } from './Resource';
+import { Resource } from './Resource';
+import { EMPTY_RESOURCE, resourceFromDetectedResource } from './ResourceImpl';
 import { ResourceDetectionConfig } from './config';
 
 /**
@@ -26,8 +26,8 @@ import { ResourceDetectionConfig } from './config';
  */
 export const detectResources = (
   config: ResourceDetectionConfig = {}
-): IResource => {
-  const resources: IResource[] = (config.detectors || []).map(d => {
+): Resource => {
+  const resources: Resource[] = (config.detectors || []).map(d => {
     try {
       const resource = resourceFromDetectedResource(d.detect(config));
       diag.debug(`${d.constructor.name} found resource.`, resource);
@@ -52,7 +52,7 @@ export const detectResources = (
  *
  * @param resources The array of {@link Resource} that should be logged. Empty entries will be ignored.
  */
-const logResources = (resources: Array<IResource>) => {
+const logResources = (resources: Array<Resource>) => {
   resources.forEach(resource => {
     // Print only populated resources
     if (Object.keys(resource.attributes).length > 0) {
