+++ packages/opentelemetry-resources/test/detectors/browser/ProcessDetector.test.ts
@@ -17,7 +17,7 @@ import * as sinon from 'sinon';
 import { processDetector } from '../../../src';
 import { describeBrowser } from '../../util';
 import { assertEmptyResource } from '../../util/resource-assertions';
-import { resourceFromDetectedResource } from '../../../src/Resource';
+import { resourceFromDetectedResource } from '../../../src/ResourceImpl';
 
 describeBrowser('processDetector() on web browser', () => {
   afterEach(() => {
