+++ packages/sdk-metrics/src/MeterProvider.ts
@@ -21,7 +21,7 @@ import {
   MeterOptions,
   createNoopMeter,
 } from '@opentelemetry/api';
-import { DEFAULT_RESOURCE, IResource } from '@opentelemetry/resources';
+import { DEFAULT_RESOURCE, Resource } from '@opentelemetry/resources';
 import { IMetricReader } from './export/MetricReader';
 import { MeterProviderSharedState } from './state/MeterProviderSharedState';
 import { MetricCollector } from './state/MetricCollector';
@@ -33,7 +33,7 @@ import { View, ViewOptions } from './view/View';
  */
 export interface MeterProviderOptions {
   /** Resource associated with metric telemetry  */
-  resource?: IResource;
+  resource?: Resource;
   views?: ViewOptions[];
   readers?: IMetricReader[];
 }
